+++ lib/multiverse/railtie.rb
@@ -38,6 +38,16 @@ module Multiverse
             ActiveRecord::Tasks::DatabaseTasks.purge ActiveRecord::Base.configurations[Multiverse.env("test")]
           end
         end
+
+        namespace :schema do
+          namespace :cache do
+            task dump: [:environment, :load_config] do
+              conn = Multiverse.record_class.connection
+              filename = File.join(ActiveRecord::Tasks::DatabaseTasks.db_dir, "schema_cache.yml")
+              ActiveRecord::Tasks::DatabaseTasks.dump_schema_cache(conn, filename)
+            end
+          end
+        end
       end
 
       namespace :multiverse do